requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
elevenlabs
textstat
tenacity

# extras needed by your endpoints
python-multipart   # for UploadFile form handling
pypdf              # lightweight PDF text extraction
httpx              # async HTTP client (useful for tests / integrations)
requests           # many SDKs depend on this